fix: StudentControllerTestにおけるバリデーションのテストをStudentDetailTestに移動
